refactor: extract release-gate shared utilities
